修正：PDF画像インポート時の型不一致エラーを修正
- pageHeight (decimal) と image.Bounds.Top (double) の計算で発生するビルドエラーを修正
- 両方の値を float にキャストしてから減算するように変更

+++ Helpers/VectorFileParser.cs
@@ -155,7 +155,7 @@ namespace FigCrafterApp.Helpers
                         {
                             ImageData = bitmap,
                             X = (float)image.Bounds.Left,
-                            Y = (float)(pageHeight - image.Bounds.Top),
+                            Y = (float)pageHeight - (float)image.Bounds.Top,
                             Width = (float)image.Bounds.Width,
                             Height = (float)image.Bounds.Height,
                             Opacity = 1.0f
